[Core] Fix sdk style projects not having item def properties
Project items in SDK style projects now have properties from item
definitions available. Changed the implementation so that instead
of trying to use the evaluated item definition properties from the
main MSBuildProject, which did not work for update items since
the MSBuildItem had a different parent MSBuildProject, now the
properties are copied across to the evaluated item if they do not
exist already.